Refactoring: move memory reference analysis to separate module
